requirements.txt
Pillow
imagehash
click
